+++ Discordian/Discordian/App.xaml.cs
@@ -75,7 +75,8 @@ namespace Discordian
 
         private ActivationService CreateActivationService()
         {
-            return new ActivationService(this, typeof(Views.MainPage), new Lazy<UIElement>(CreateShell));
+            //return new ActivationService(this, typeof(Views.MainPage), new Lazy<UIElement>(CreateShell));
+            return new ActivationService(this, typeof(Views.LoginPage));
         }
 
         private UIElement CreateShell()
