+++ Discordian/Discordian/Views/LoginPage.xaml
@@ -1,0 +1,42 @@
+﻿<Page
+    x:Class="Discordian.Views.LoginPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Discordian.Views"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Grid>
+        <Grid.Background>
+            <ImageBrush Stretch="Fill" ImageSource="/Assets/Background/login.jpg"/>
+        </Grid.Background>
+        <StackPanel
+            x:Name="LoginContentDialog"
+            Width="300"
+            Height="404"
+            HorizontalAlignment="Center"
+            VerticalAlignment="Center"
+            Background="White"
+            CornerRadius="5">
+            <Grid>
+                <StackPanel Margin="20">
+                    <TextBlock HorizontalAlignment="Center" FontSize="30" CharacterSpacing="100" FontFamily="/Assets/Fonts/AquireLight-YzE0o.otf#Aquire Light" FontWeight="Normal">Login</TextBlock>
+                    <TextBox x:Name="EmailTextBox" Header="Email" PlaceholderText="Discordian email..." HorizontalAlignment="Stretch" Margin="0,0,0,20">
+                    </TextBox>
+                    <PasswordBox x:Name="PasswordTextBox" Margin="0,0,0,10" Header="Password" PlaceholderText="Discordian password..." HorizontalAlignment="Stretch"/>
+                    <TextBlock x:Name="AccountValidationErrorMessage" HorizontalAlignment="Center" Margin="10" Visibility="Collapsed" Foreground="Red"></TextBlock>
+
+                    <StackPanel>
+                        <Grid>
+                            <Button  Content="Login" x:Name="loginBtn" Click="LoginButton_Click" Margin="10" Width="100"  HorizontalAlignment="Center"/>
+                        </Grid>
+                    </StackPanel>
+                </StackPanel>
+                <muxc:ProgressRing x:Name="LoginSpinner" IsActive="False" Width="80" Height="80" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" />
+            </Grid>
+        </StackPanel>
+    </Grid>
+</Page>
