+++ Discordian/Discordian/Views/MainPage.xaml.cs
@@ -20,8 +20,6 @@ namespace Discordian.Views
 {
     public sealed partial class MainPage : Page
     {
-        private readonly string AuthorizeURL = "https://dgrozev99.wixsite.com/my-site/_functions-dev/authorize";
-        private static bool isLoggedIn = false;
         public MainViewModel ViewModel { get; } = new MainViewModel();
 
         public MainPage()
@@ -33,45 +31,6 @@ namespace Discordian.Views
 
         private async void MainPage_Loaded(object sender, RoutedEventArgs e)
         {
-            if (!isLoggedIn)
-            {
-                await this.LoginContentDialog.ShowAsync();
-            }
-            else
-            {
-                await this.LoadStatisticsDataAsync();
-            }
-        }
-
-        private void OnCloseRequest(object sender, SystemNavigationCloseRequestedPreviewEventArgs e)
-        {
-            DiscordianBotConsoleClient.StopAllBots();
-        }
-
-        private async void LoginButton_Click(object sender, RoutedEventArgs e)
-        {
-            LoginSpinner.IsActive = true;
-
-            var isSubscribed = await this.CheckIfUserIsSubscribedAsync();
-
-            LoginSpinner.IsActive = false;
-
-            if (isSubscribed)
-            {
-                await this.LoadStatisticsDataAsync();
-
-                isLoggedIn = true;
-            }
-            else
-            {
-                this.AccountValidationErrorMessage.Text = "Please renew your subscription!";
-            }
-        }
-
-        private async Task LoadStatisticsDataAsync()
-        {
-            this.LoginContentDialog.Hide();
-
             if (App.Connection == null)
             {
                 await FullTrustProcessLauncher.LaunchFullTrustProcessForCurrentAppAsync("Id");
@@ -122,30 +81,9 @@ namespace Discordian.Views
             ProgressSpinnerForMentionsPerBot.IsActive = false;
         }
 
-        private async Task<bool> CheckIfUserIsSubscribedAsync()
+        private void OnCloseRequest(object sender, SystemNavigationCloseRequestedPreviewEventArgs e)
         {
-            var email = this.EmailTextBox.Text;
-            var password = this.PasswordTextBox.Password;
-
-            try
-            {
-                var httpClient = new HttpClient();
-                httpClient.DefaultRequestHeaders.Add("user-agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/71.0.3578.98 Safari/537.36");
-                var httpRequestMessage = new HttpRequestMessage();
-                var content = new HttpStringContent(await Json.StringifyAsync(new User { email = email, password = password }), Windows.Storage.Streams.UnicodeEncoding.Utf8, "application/json");
-                var url = new Uri(AuthorizeURL);
-
-                var responseMessage = await httpClient.PostAsync(url, content);
-                var responseBody = await responseMessage.Content.ReadAsStringAsync();
-                var data = await Json.ToObjectAsync<AuthenticationResponse>(responseBody);
-
-                return data.IsSubscribed;
-            }
-            catch (Exception)
-            {
-            }
-
-            return false;
+            DiscordianBotConsoleClient.StopAllBots();
         }
     }
 }
